Moved Packages and Started Testing
Moved the classes to a new routeFinder package, and created tests.
