--- routeFinder/.gitignore
@@ -1,6 +1,0 @@
-/AStarSearch.class
-/Main.class
-/Map$Point.class
-/Map$Street.class
-/Map.class
-/SearchPoint.class
