+++ routeFinder/Main.java
@@ -1,26 +1,26 @@
 package routeFinder;
 import java.util.ArrayList;
-
 import processing.core.*;
-import routeFinder.Map.Point;
 
 /*
  * Original implementation by Gary Dahl
  */
 public class Main extends PApplet
 {
-	AStarSearch search; // your implementation of A* search
-	Map map; // map to search for path between start and end points of
-	boolean enterPressed; // press enter to watch entire search until solution
-	boolean spaceWasDown; // press space repeatedly to step through search
+	public AStarSearch search; // your implementation of A* search
+	public Map map; // map to search for path between start and end points of
+	public boolean enterPressed; // press enter to watch entire search until solution
+	public boolean spaceWasDown; // press space repeatedly to step through search
 
+	public String mapFileName = "map.osm";
+	
 	// initialize window
 	public void settings() { size(800, 600); }
 	
 	// load map, and initialize fields along with processing modes
 	public void setup()
 	{		
-		map = new Map("map.osm",this);
+		map = new Map(mapFileName,this);
 		search = null;
 		spaceWasDown = false;
 		enterPressed = false;
@@ -83,7 +83,7 @@ public class Main extends PApplet
 		updateGUI(); // allow user to drag around end points		
 	}
 	
-	public void colorPoints(ArrayList<Map.Point> points, boolean isFrontier)
+	public void colorPoints(ArrayList<Point> points, boolean isFrontier)
 	{
 		if(isFrontier) 
 		{
@@ -98,11 +98,11 @@ public class Main extends PApplet
 			text("EXPLORED: "+points.size(),width/2,32);
 		}
 		
-		for(Map.Point p : points)
+		for(Point p : points)
 			ellipse(p.x,p.y,4,4);
 	}
 	
-	public void drawSolution(ArrayList<Map.Point> points)
+	public void drawSolution(ArrayList<Point> points)
 	{
 		stroke(255); // draw white lines between points
 		strokeWeight(2);
