+++ tests/.gitignore
@@ -1,1 +1,3 @@
 /SearchPointTest.class
+/OSMRouteFinderEndToEndTest.class
+/ApplicationRunner.class
