--- tests/SearchPointTest.java
@@ -1,21 +1,0 @@
-package tests;
-
-import static org.junit.Assert.*;
-
-import org.junit.Before;
-import org.junit.Test;
-
-public class SearchPointTest {
-
-	
-	@Before
-	public void setUp() throws Exception {
-		
-	}
-
-	@Test
-	public void test() {
-		fail("Not yet implemented");
-	}
-
-}
